moved disableOnWheelInput to helpers
