+++ src/components/GrantRequest/CategorySection/AcceleratorSection.tsx
@@ -5,13 +5,13 @@ import useEditor, { assert, createValidator } from 'decentraland-gatsby/dist/hoo
 import useFormatMessage from 'decentraland-gatsby/dist/hooks/useFormatMessage'
 import { Field } from 'decentraland-ui/dist/components/Field/Field'
 
-import { disableOnWheelInput } from '../../../entities/Common/utils'
 import {
   AcceleratorQuestions,
   AcceleratorQuestionsSchema,
   GrantRequestCategoryAssessment,
 } from '../../../entities/Grant/types'
 import { asNumber } from '../../../entities/Proposal/utils'
+import { disableOnWheelInput } from '../../../helpers'
 import { useGrantCategoryEditor } from '../../../hooks/useGrantCategoryEditor'
 import Label from '../../Common/Label'
 import { ContentSection } from '../../Layout/ContentLayout'
