+++ src/components/Proposal/Submit/ProposalSubmitPoiPage.tsx
@@ -13,9 +13,9 @@ import { Button } from 'decentraland-ui/dist/components/Button/Button'
 import { Header } from 'decentraland-ui/dist/components/Header/Header'
 
 import { Governance } from '../../../clients/Governance'
-import { disableOnWheelInput } from '../../../entities/Common/utils'
 import { PoiType, getPoiTypeAction, newProposalPOIScheme } from '../../../entities/Proposal/types'
 import { asNumber, isAlreadyPointOfInterest, isValidPointOfInterest } from '../../../entities/Proposal/utils'
+import { disableOnWheelInput } from '../../../helpers'
 import loader from '../../../modules/loader'
 import locations from '../../../modules/locations'
 import Field from '../../Common/Form/Field'
