+++ src/pages/submit/linked-wearables.tsx
@@ -25,9 +25,9 @@ import ContentLayout, { ContentSection } from '../../components/Layout/ContentLa
 import LoadingView from '../../components/Layout/LoadingView'
 import CoAuthors from '../../components/Proposal/Submit/CoAuthor/CoAuthors'
 import LogIn from '../../components/User/LogIn'
-import { disableOnWheelInput } from '../../entities/Common/utils'
 import { newProposalLinkedWearablesScheme } from '../../entities/Proposal/types'
 import { asNumber, isValidImage, userModifiedForm } from '../../entities/Proposal/utils'
+import { disableOnWheelInput } from '../../helpers'
 import { isHttpsURL } from '../../helpers'
 import loader from '../../modules/loader'
 import locations from '../../modules/locations'
